move proto related code into single file, documentation
